+++ src/__tests__/runner.test.js
@@ -429,6 +429,13 @@ describe('Runner', () => {
       expect(close).toBeCalled();
     });
 
+    it('emits debuggerProcessExit when process close', () => {
+      const close = jest.fn();
+      runner.on('debuggerProcessExit', close);
+      fakeProcess.emit('close');
+      expect(close).toBeCalled();
+    });
+
     it('should start jest process after killing the old process', () => {
       runner.closeProcess();
       runner.start();